Improve Zooming and Panning in Lightbox
Allow setting a minumum Zoom-Out scale.

Allow resetting the view to the initial view when ZoomMin has
been reached.  This recenters the image.

Allow for RelativePinchtoZoom where successive pinching results
in additional zooming rather than starting over each time.

Define ZoomMax as where 1 pixel in the image equals 1 pixel on
the display.

Allow defining a ZoomOut scaling factor for pinch to zoom.
Zooming by default is slow because its scale is [0,1] where as
zooming in is [1,inf]. A scalling factor of 3 seems to create
the expected zoom out functionality.

Limit panning so that the edge of the image is not allowed to
go past the center of the viewer.